pyproject.toml

[project]
authors = [
  {name = "Matias Iturburu", email = "matias.iturburu@bigpear.com.ar"},
]
description = ""
name = "reservas"
readme = "README.md"

license = {text = "MIT"}

dependencies = [
  "django",
  "django-rest-framework",
  "django-environ",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
  "ruff",
  "coverage",
  "flake8",
  "flake8-pyproject",
  "ipdb",
  "pre-commit",
  "pytest",
  "pytest-django",
  "mypy",
  "django-stubs",
  "djangorestframework-stubs",
]
docs = ["docutils>=0.3", "pack ==1.1, ==1.3"]

[project.scripts]
reservas = "reservas.manage:main"

[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools", "setuptools-scm"]

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.isort]
known_first_party = [
  "reservas",
]

[tool.flake8]
exclude = ['**/migrations/**']
extend-ignore = ["E203"]
ignore = ["E501", "F403", "E402", "E203", "W503", "Q000"]
max-complexity = 12
max-line-length = 120
statistics = true

[tool.ruff]
exclude = ["test/data"]
line-length = 120

[tool.ruff.lint]
# max-args = 20
# good-names = ["i", "j", "k", "ex", "_", "e", "db", "df", "ms", "ts"]
# ignore = ["E114", "C0115", "C0116","E0611", "R0902", "R0903", "R0913"]
ignore = ["D100", "D101", "D102", "D103"]

[tool.ruff.lint.flake8-unused-arguments]
ignore-variadic-names = true

[tool.mypy]
plugins = [
  "mypy_django_plugin.main",
  "mypy_drf_plugin.main",
]
python_version = "3.11"

[tool.django-stubs]
django_settings_module = "reservas.conf.settings"

[tool.pytest.ini_options]
addopts = [
  "-p no:cacheprovider",
  "-ra",
  "--strict-markers",
  "--verbose",
  "--cov-config=.coveragerc",
  "--cov=src", #"--cov-report=term",
  "--cov-report=html:docs/coverage",
  "--cov-report=xml:docs/coverage/coverage.xml",
  "--junitxml=docs/test/junit/test-results.xml",
  "--verbose",
  "--cache-clear",
  "-m not slow",
]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
python_files = ["test_*.py"]
testpaths = [
  "tests",
]

[tool.coverage.run]
omit = []

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "@abstractmethod",
]

[tool.pytest]
DJANGO_SETTINGS_MODULE = "reservas.conf.settings"
python_files = "tests.py test_*.py *_tests.py"
